Use a single record + type instead of separate functions. Move registry to data.
